refactor: 이메일 타입 패키지 이동
